+++ bundles/framework/bundle/mapanalysis/locale/en.js
@@ -1,9 +1,15 @@
 Oskari.registerLocalization({
+
     "lang" : "en",
     "key" : "MapAnalysis",
     "value" : {
         "title" : "MapAnalysis",
         "desc" : "",
-        "object-data": "Object data"
+        "object-data": "Object data",
+         "layer" : {
+                "organization" : "Analysis",
+                "inspire" : "Analysis"
+
+            }
     }
 });