+++ bundles/framework/bundle/mapanalysis/locale/fi.js
@@ -1,9 +1,15 @@
 Oskari.registerLocalization({
+
     "lang" : "fi",
     "key" : "MapAnalysis",
     "value" : {
         "title" : "MapAnalysis",
         "desc" : "",
-        "object-data": "Kohdetiedot"
+        "object-data": "Kohdetiedot",
+        "layer" : {
+                "organization" : "Analyysit",
+                "inspire" : "Analyysitasot"
+
+            }
     }
 });