+++ bundles/framework/bundle/publisher/instance.js
@@ -179,6 +179,30 @@ function() {
                 
             dialog.show(loc['published'].title, content, [okBtn]);
             this.setPublishMode(false);
+        },
+        /**
+         * @method Publisher.ToolStyleChangedEvent
+         */
+        'Publisher.ToolStyleChangedEvent': function(event) {
+            if (!this.publisher) return;
+
+            this.publisher.changeToolStyles(event.getStyle());
+        },
+        /**
+         * @method Publisher.ColourSchemeChangedEvent
+         */
+        'Publisher.ColourSchemeChangedEvent': function(event) {
+            if (!this.publisher) return;
+
+            this.publisher.changeColourScheme(event.getColourScheme());
+        },
+        /**
+         * @method Publisher.FontChangedEvent
+         */
+        'Publisher.FontChangedEvent': function(event) {
+            if (!this.publisher) return;
+
+            this.publisher.changeFont(event.getFont());
         }
     },