+++ bundles/framework/bundle/publisher/locale/fi.js
@@ -60,6 +60,43 @@ Oskari.registerLocalization(
                     "tooltip": "Näytä karttaan liittyvä taulukko.",
                     "grid" : "Näytä tilastotaulukko"
                 },
+                "layout": {
+                    "label": "Ulkoasu",
+                    "fields": {
+                        "colours": {
+                            "label": "Värimaailma",
+                            "placeholder": "Valitse värimaailma",
+                            "buttonLabel": "Vaihda",
+                            "light_grey": "Vaalean harmaa",
+                            "dark_grey": "Tumman harmaa",
+                            "blue": "Sininen",
+                            "red": "Punainen",
+                            "green": "Vihreä",
+                            "yellow": "Keltainen"
+                        },
+                        "fonts": {
+                            "label": "Valitse fontti"
+                        },
+                        "toolStyles": {
+                            "label": "Työkalujen tyyli",
+                            "rounded-dark": "Pyöristetty (tumma)",
+                            "rounded-light": "Pyöristetty (vaalea)",
+                            "sharp-dark": "Kulmikas (tumma)",
+                            "sharp-light": "Kulmikas (vaalea)",
+                            "3d-dark": "3D (tumma)",
+                            "3d-light": "3D (vaalea)"
+                        }
+                    },
+                    "popup": {
+                        "title": "Värimaailman valinta",
+                        "close": "Sulje",
+                        "gfiDialog": {
+                            "title": "Kohteen tiedot",
+                            "featureName": "Esikatselu",
+                            "featureDesc": "Värimaailman valinta vaikuttaa ainoastaan kohdetietolaatikon väritykseen"
+                        }
+                    }
+                },
                 "layers": {
                     "label": "Karttatasot",
                     "defaultLayer": "(Oletusvalinta)",