+++ bundles/integration/bundle/admin-layerselector/View.js
@@ -16,8 +16,7 @@
  *
  *
  */
-Oskari.clazz.define('Oskari.integration.bundle.admin-layerselector.View', function() {
-}, {
+Oskari.clazz.define('Oskari.integration.bundle.admin-layerselector.View', function () {}, {
 
     /**
      * @property eventHandlers
@@ -25,27 +24,29 @@ Oskari.clazz.define('Oskari.integration.bundle.admin-layerselector.View', functi
      * These will be registered/unregistered automagically
      *
      */
-    "eventHandlers" : {
-        "MapLayerVisibilityChangedEvent" : function(event) {
-            
-        },
-        "AfterMapMoveEvent" : function(event) {
+    "eventHandlers": {
+        "MapLayerVisibilityChangedEvent": function (event) {
 
         },
-        'MapLayerEvent' : function(event) {
+        "AfterMapMoveEvent": function (event) {
+
+        },
+        'MapLayerEvent': function (event) {
             // trigger after interval since events are being spammed by backendstatus
             // this way browser doesn't crash
             var interval = 500,
                 me = this;
-            if(me._previousLayerUpdateTimer) {
+            if (me._previousLayerUpdateTimer) {
                 clearTimeout(me._previousLayerUpdateTimer);
                 me._previousLayerUpdateTimer = null;
             }
-            me._previousLayerUpdateTimer = setTimeout(function() {
+            me._previousLayerUpdateTimer = setTimeout(function () {
                 me._layerUpdateHandler();
             }, interval);
 
-            jQuery("body").css({cursor: "auto"});
+            jQuery("body").css({
+                cursor: "auto"
+            });
         }
     },
     /**
@@ -53,17 +54,17 @@ Oskari.clazz.define('Oskari.integration.bundle.admin-layerselector.View', functi
      * @private
      * Updates layers listing after layers has been changed/MapLayerEvent has been received.
      */
-    _layerUpdateHandler : function() {
-		//console.log("admin-layerselector/View.js:_layerUpdateHandler");
+    _layerUpdateHandler: function () {
+        //console.log("admin-layerselector/View.js:_layerUpdateHandler");
         // TODO! currently update, add and initial additions execute
         // the same code. This needs to be updated when mapLayerService
         // can handle updates better. 
         // (updates everything instead of layer.name)
         var sandbox = this.getSandbox(),
-        // populate layer list
+            // populate layer list
             mapLayerService = sandbox.getService('Oskari.mapframework.service.MapLayerService'),
             layers = mapLayerService.getAllLayers();
-        if(this.view != null){
+        if (this.view != null) {
             this.view.addToCollection(layers);
         } else {
             this.layers = layers;
@@ -76,17 +77,17 @@ Oskari.clazz.define('Oskari.integration.bundle.admin-layerselector.View', functi
         }
         */
     },
-    
+
     /**
      * @property requirementsConfig
      *
      * requirejs requirements config to fix paths
      *
      */
-    "requirementsConfig" : {
-        "waitSeconds" : 15,
-        "paths" : {
-            '_bundle' : '../../../Oskari/bundles/integration/bundle/admin-layerselector'
+    "requirementsConfig": {
+        "waitSeconds": 15,
+        "paths": {
+            '_bundle': '../../../Oskari/bundles/integration/bundle/admin-layerselector'
         }
     },
 
@@ -96,7 +97,7 @@ Oskari.clazz.define('Oskari.integration.bundle.admin-layerselector.View', functi
      * This is called when flyout is ready and something needs to be executed
      * before Backbone is rendered
      */
-    init : function() {
+    init: function () {
         // if something needs to be initiated.
     },
 
@@ -106,31 +107,33 @@ Oskari.clazz.define('Oskari.integration.bundle.admin-layerselector.View', functi
      * Called with requirements from above as arguments to method in
      * defined order.
      */
-    "render" : function() {
+    "render": function () {
         var me = this,
             container = me.getEl();
         // admin-layerselector is rendered under this container
         container.addClass("admin-layerselector");
         // backbone will fire adminAction events if they need to be 
         // passed to other bundles
-        container.on("adminAction", {me: me}, me.handleAction);
+        container.on("adminAction", {
+            me: me
+        }, me.handleAction);
 
         var locale = me.getLocalization(),
-            confRequirementsConfig = 
-            (this.getConfiguration()||{}).requirementsConfig,
-            requirementsConfig = 
-            confRequirementsConfig||this.requirementsConfig;
+            confRequirementsConfig =
+                (this.getConfiguration() || {}).requirementsConfig,
+            requirementsConfig =
+                confRequirementsConfig || this.requirementsConfig;
 
         require.config(requirementsConfig);
-        require(["_bundle/views/layerSelectorView"], function(LayerSelectorView) {
+        require(["_bundle/views/layerSelectorView"], function (LayerSelectorView) {
 
             // Finally, we kick things off by creating the **App**.
             // We need to pass container element for the view and
             // instance. 
             me.view = new LayerSelectorView({
-                el : container,
-                instance : me.instance,
-                locale : me.locale
+                el: container,
+                instance: me.instance,
+                locale: me.locale
             });
             // If call for layers is ready before backbone is created,
             // we'll instantiate our view with that data 
@@ -145,7 +148,7 @@ Oskari.clazz.define('Oskari.integration.bundle.admin-layerselector.View', functi
      * This is called when backbone fires an event that needs to be passed
      * to other bundles. Event should contain a member called *command*
      */
-    handleAction: function(e) {
+    handleAction: function (e) {
         e.stopPropagation();
         var me = e.data.me,
             sandbox = me.getSandbox(),
@@ -155,7 +158,7 @@ Oskari.clazz.define('Oskari.integration.bundle.admin-layerselector.View', functi
          * NORMAL LAYERS *
          *****************
          */
-        if(e.command == "removeLayer") {
+        if (e.command == "removeLayer") {
             // remove layer from mapLayerService
             if (e.baseLayerId) {
                 // If this is a sublayer, remove it from its parent's sublayer array
@@ -165,7 +168,7 @@ Oskari.clazz.define('Oskari.integration.bundle.admin-layerselector.View', functi
                 // otherwise just remove it from map layer service.
                 mapLayerService.removeLayer(e.modelId);
             }
-        } else if(e.command == "addLayer") {
+        } else if (e.command == "addLayer") {
             // add layer into mapLayerService
             e.layerData.name = e.layerData.admin.name[Oskari.getDefaultLanguage()];
             var mapLayer = mapLayerService.createMapLayer(e.layerData);
@@ -177,13 +180,13 @@ Oskari.clazz.define('Oskari.integration.bundle.admin-layerselector.View', functi
                 mapLayerService.addSubLayer(parentLayerId, mapLayer);
             } else {
                 // Otherwise just add it to the map layer service.
-                if(mapLayerService._reservedLayerIds[mapLayer.getId()] !== true) {
+                if (mapLayerService._reservedLayerIds[mapLayer.getId()] !== true) {
                     mapLayerService.addLayer(mapLayer);
                 }
             }
-        } else if(e.command == "editLayer") {
+        } else if (e.command == "editLayer") {
             // update layer info
-			//console.log("Editing layer");
+            //console.log("Editing layer");
             //console.log(e.layerData.admin);
             e.layerData.name = e.layerData.admin.name[Oskari.getDefaultLanguage()]; //TODO this should be in mapLayerService
             mapLayerService.updateLayer(e.layerData.id, e.layerData);
@@ -195,23 +198,22 @@ Oskari.clazz.define('Oskari.integration.bundle.admin-layerselector.View', functi
          ************************
          */
         // load the map layers again
-        else if(e.command == "addGroup") {
+        else if (e.command == "addGroup") {
             mapLayerService.loadAllLayersAjax();
         }
         // Remove the base/group layer from mapLayerService
         // and load it again from backend, since we edited
         // the layer class and the changes will not be
         // reflected to the corresponding map layer directly.
-        else if(e.command == "editGroup") {
+        else if (e.command == "editGroup") {
             mapLayerService.removeLayer(e.id, true);
             mapLayerService.loadAllLayersAjax();
-        }
-        else if(e.command == "deleteGroup") {
+        } else if (e.command == "deleteGroup") {
             mapLayerService.removeLayer(e.id);
         }
     }
 
 
 }, {
-    "extend" : ["Oskari.integration.bundle.bb.View"]
+    "extend": ["Oskari.integration.bundle.bb.View"]
 });